Abstracted template_check_generator for cleaner inheritence
