+++ gemd/entity/object/has_spec.py
@@ -1,6 +1,7 @@
 """For entities that have specs."""
 from gemd.entity.has_dependencies import HasDependencies
 from gemd.entity.object.has_template import HasTemplate
+from gemd.entity.object.has_template_check_generator import HasTemplateCheckGenerator
 from gemd.entity.template.base_template import BaseTemplate
 from gemd.entity.link_by_uid import LinkByUID
 
@@ -10,7 +11,7 @@ from typing import Optional, Union, Set, Type, Callable, TypeVar
 T = TypeVar('T')
 
 
-class HasSpec(HasDependencies):
+class HasSpec(HasTemplateCheckGenerator, HasDependencies):
     """Mix-in trait for objects that can be assigned specs.
 
     Parameters
@@ -52,12 +53,6 @@ class HasSpec(HasDependencies):
             return self.spec.template
         else:
             return None
-
-    def _generate_template_check(self,
-                                 validate: Callable[[Union["HasSpec", "HasTemplate"], T], bool]
-                                 ) -> Callable[[T], T]:
-        """Generate a closure for the object and the validation routine (from HasTemplate)."""
-        return HasTemplate._generate_template_check(self, validate)
 
     def _local_dependencies(self) -> Set[Union["BaseEntity", "LinkByUID"]]:
         """Return a set of all immediate dependencies (no recursion)."""
